+++ src/store/actions.ts
@@ -4,10 +4,9 @@ import { createAction } from 'typesafe-actions';
 
 import { TX_DEFAULTS } from '../common/constants';
 import { getContractWrappers } from '../services/contract_wrappers';
+import { cancelSignedOrder, getAllOrdersAsUIOrders, getUserOrdersAsUIOrders } from '../services/orders';
+import { getTokenBalance, tokenToTokenBalance } from '../services/tokens';
 import { getWeb3Wrapper, getWeb3WrapperOrThrow } from '../services/web3_wrapper';
-import { cancelSignedOrder } from '../util/cancel_order';
-import { getAllOrdersAsUIOrders, getUserOrdersAsUIOrders } from '../util/get_orders';
-import { getTokenBalance, tokenToTokenBalance } from '../util/get_token_balance';
 import { getKnownTokens, getTokenBySymbol, getWethToken } from '../util/known_tokens';
 import { BlockchainState, RelayerState, Token, TokenBalance, UIOrder, Web3State } from '../util/types';
